+++ internal/provider/bucket_cors_configuration_test.go
@@ -92,7 +92,7 @@ func TestAccBucketCORSConfigurationResource(t *testing.T) {
 		),
 	}
 
-	resource.Test(t, resource.TestCase{
+	resource.ParallelTest(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{