+++ CHANGELOG.md
@@ -5,6 +5,13 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.3.0]
+
+### Added:
+
+- objsto_bucket_versioning resource for configuring bucket versioning settings.
+- objsto_object: `version_id` attribute to store the version ID of the object.
+
 ## [0.2.2]
 
 ### Fixed
@@ -36,7 +43,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 - Minimal implementation of bucket, bucket policy, bucket lifecycle configuration, and object resources.
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v0.1.1...v0.2.0